+++ src/app/core/layout/layout.module.ts
@@ -1,0 +1,21 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { RouterModule } from '@angular/router';
+
+import { SidebarComponent } from './sidebar/sidebar.component';
+import { MaterialModule } from 'src/app/shared/material.module';
+
+@NgModule({
+  declarations: [
+    SidebarComponent
+  ],
+  imports: [
+    CommonModule,
+    MaterialModule,
+    RouterModule,
+  ],
+  exports: [
+    SidebarComponent,
+  ]
+})
+export class LayoutModule { }
